progress torwards correct demos
Ordering matters a lot for our P_Random()
function with global side effects.

First issue is with code that uses it like this:

P_Random() - P_Random()

compared to gcc this had two issues, the ordering
was different and becuase this function returned a
uchar we were not allowing it to become negative.

So we steal P_Random2() from our doom, and add our
own P_Random3 to cover the large amount of P_Random() - 128
that would also have this sign issue.

The second issue is in how inlined calls are ordered example:

fn(P_Rand() + 1, P_Random() + 2, P_Random() + 3);

we order how you read it, left to right. gcc does it right to left.
So in relevant places we use locals and reverse the ordering of how
these are called.

This fixes these issues for callers of P_SpawnMobj, which seemed
to be a decent chunk of the offenders.

This has impacted demos, they are arguably 'closer' to accurate.
demo1 still goes out of sync due to a hit connecting in our demo
playback but not in the original. This causes/misses a hitsplat and
that throws off the entire prng. Likely due to similar issues in the
enemy movement logic.